refactor: Extract VideoToggleButton
