+++ src/components/TalkInfo/TalkInfo.tsx
@@ -2,10 +2,9 @@ import React, { useEffect, useState } from 'react'
 import * as Styled from './styled'
 import { Event, Talk } from '../../generated/dreamkast-api.generated'
 import { useGetApiV1TracksByTrackIdViewerCountQuery } from '../../generated/dreamkast-api.generated'
-import { Button } from '@material-ui/core'
-import { useDispatch, useSelector } from 'react-redux'
-import { setShowVideo, settingsSelector } from '../../store/settings'
-import { Videocam, VideocamOff } from '@material-ui/icons'
+import { useSelector } from 'react-redux'
+import { settingsSelector } from '../../store/settings'
+import { VideoToggleButton } from '../common/VideoToggleButton'
 
 type Props = {
   event?: Event
@@ -20,7 +19,6 @@ export const TalkInfo: React.FC<Props> = ({
   selectedTrackName,
   selectedTrackId,
 }) => {
-  const dispatch = useDispatch()
   const [viewerCount, setViewerCount] = useState<string>()
   const settings = useSelector(settingsSelector)
 
@@ -54,18 +52,7 @@ export const TalkInfo: React.FC<Props> = ({
   return (
     <Styled.Container>
       {selectedTalk?.onAir && <Styled.Live>LIVE 👥 {viewerCount}</Styled.Live>}
-      {settings.profile.isAttendOffline && (
-        <Button
-          variant="contained"
-          size="small"
-          color="primary"
-          disableElevation
-          startIcon={settings.showVideo ? <VideocamOff /> : <Videocam />}
-          onClick={() => dispatch(setShowVideo(!settings.showVideo))}
-        >
-          {settings.showVideo ? 'Stop Video' : 'Show Video'}
-        </Button>
-      )}
+      {settings.profile.isAttendOffline && <VideoToggleButton />}
       <Styled.Title>{selectedTalk?.title}</Styled.Title>
       <Styled.SpeakerContainer>
         <Styled.Speaker>
